modifications required to compile on arch linux

+++ models/color.cpp
@@ -181,21 +181,21 @@ namespace stg {
         hue = hue * 360 / 60;
 
         auto chroma = (1 - std::abs(2 * lightness - 1)) * saturation;
-        auto x = chroma * (1 - std::abs(std::fmodf(hue, 2) - 1));
+        auto x = chroma * (1 - std::abs(fmodf(hue, 2) - 1));
 
         auto rgb = std::array<float, 3>();
 
-        if (std::ceilf(hue) == 1) {
+        if (std::ceil(hue) == 1) {
             rgb = {chroma, x, 0};
-        } else if (std::ceilf(hue) == 2) {
+        } else if (std::ceil(hue) == 2) {
             rgb = {x, chroma, 0};
-        } else if (std::ceilf(hue) == 3) {
+        } else if (std::ceil(hue) == 3) {
             rgb = {0, chroma, x};
-        } else if (std::ceilf(hue) == 4) {
+        } else if (std::ceil(hue) == 4) {
             rgb = {0, x, chroma};
-        } else if (std::ceilf(hue) == 5) {
+        } else if (std::ceil(hue) == 5) {
             rgb = {x, 0, chroma};
-        } else if (std::ceilf(hue) == 6) {
+        } else if (std::ceil(hue) == 6) {
             rgb = {chroma, 0, x};
         }
